+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## 2.0.0
 
+- Code made sound to support Dart strong mode.
+- Removed arbitrary properties from Request and Session: use subtypes instead.
 - Changed default bindAddress from LOOPBACK_IP_V6 to LOOPBACK_IP_V4.
 - Added convenience methods for registering PUT, PATCH and DELETE handlers.
-- Code made sound to support Dart strong mode.
 
 ## 1.0.5